feat(RoundSummary): first version of RoundSummary
