+++ src/components/round-summary/avatar/assets/index.ts
@@ -1,0 +1,11 @@
+import enemy1 from "./enemy1.svg";
+import enemy2 from "./enemy2.svg";
+import enemy3 from "./enemy3.svg";
+import player from "./player.svg";
+
+export const avatarAssets = {
+  player,
+  enemy1,
+  enemy2,
+  enemy3,
+};
