+++ src/components/round-summary/player-card.module.css
@@ -1,0 +1,67 @@
+.playerCard {
+  display: flex;
+  flex-direction: row;
+  isolation: isolate;
+
+  width: 20vh;
+  height: 8vh;
+  padding: 1vh;
+  background: var(--white);
+
+  border: 2px solid var(--light-grey);
+  box-shadow: 0vh 0vh 0.5vh var(--shadow-grey);
+  border-radius: 6px;
+
+  cursor: crosshair; /* + */
+}
+
+.playerCard.active {
+  z-index: 5 !important; /* FIXME */
+
+  transform: translateX(2vh);
+
+  &.player {
+    border: 2px solid var(--player-primary);
+    box-shadow: 0vh 0vh 0.5vh var(--player-secondary);
+  }
+
+  &.enemy1 {
+    border: 2px solid var(--enemy1-primary);
+    box-shadow: 0vh 0vh 0.5vh var(--enemy1-secondary);
+  }
+
+  &.enemy2 {
+    border: 2px solid var(--enemy2-primary);
+    box-shadow: 0vh 0vh 0.5vh var(--enemy2-secondary);
+  }
+
+  &.enemy3 {
+    border: 2px solid var(--enemy3-primary);
+    box-shadow: 0vh 0vh 0.5vh var(--enemy3-secondary);
+  }
+}
+
+.playerCard.clickable {
+  cursor: pointer;
+}
+
+.playerPoints {
+  flex: 2;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  align-items: center;
+}
+
+.points {
+  font-size: 1.5rem;
+  font-weight: bold;
+  color: var(--darkest-grey);
+}
+
+.extraPoints {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  background: var(--lightest-grey);
+}
