--- src/components/round-summary/player-summary.module.css
@@ -1,16 +1,0 @@
-.playerSummary {
-  --size: 80px;
-  width: var(--size);
-  height: var(--size);
-
-  display: flex;
-  flex-direction: column;
-
-  margin-bottom: 20px;
-  border-radius: 10%;
-  border: 2px solid var(--base-grey);
-}
-
-.playerSummary.clickable {
-  cursor: pointer;
-}
